backend/requirements.txt
django<4
djangorestframework==3.14.0
